\name{redisBRPOPLPUSH}
\alias{redisBRPOPLPUSH}
\title{Pop a value from a list, push it to another list and return it; or block until one is available}
\usage{
  redisBRPOPLPUSH(source, destination, timeout, Rc = NULL)
}
\description{
  Pop a value from a list, push it to another list and
  return it; or block until one is available
}

