\name{redisZINTERSTORE}
\alias{redisZINTERSTORE}
\title{Intersect multiple sorted sets and store the resulting sorted set in a new key}
\usage{
  redisZINTERSTORE(destination, numkeys, key, Rc = NULL)
}
\description{
  Intersect multiple sorted sets and store the resulting
  sorted set in a new key
}

