\name{redisZREVRANGE}
\alias{redisZREVRANGE}
\title{Return a range of members in a sorted set, by index, with scores ordered from high to low}
\usage{
  redisZREVRANGE(key, start, stop, Rc)
}
\description{
  Return a range of members in a sorted set, by index, with
  scores ordered from high to low
}

