\name{redisZREVRANGEBYSCORE}
\alias{redisZREVRANGEBYSCORE}
\title{Return a range of members in a sorted set, by score, with scores ordered from high to low}
\usage{
  redisZREVRANGEBYSCORE(key, max, min, Rc)
}
\description{
  Return a range of members in a sorted set, by score, with
  scores ordered from high to low
}

